Moved initialization steps into their own functions
